requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.2
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
dlib==19.24.0
easydict==1.10
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.1.21
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
imageio==2.26.0
insightface==0.2.1
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.10
lazy_loader==0.1
lpips==0.1.4
matplotlib==3.6.1
more-itertools==9.1.0
mpmath==1.2.1
networkx==3.0
ninja==1.10.2
numpy==1.22.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.12.0
onnxruntime==1.13.1
onnxruntime-gpu==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==23.0
Pillow==9.0.1
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==0.2.1
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.2.28
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
zipp==1.0.0
